Extract ReadyToPlayStateViewTests class
