+++ AudioBarCore/ReadyToPlayStateTests.swift
@@ -164,288 +164,3 @@ extension ReadyToPlayStateTests {
     }
 
 }
-
-extension ReadyToPlayStateTests {
-
-    func testViewPlayPauseButtonImage1() {
-        state.isPlaying = true
-        let view = state.present() as! AudioBarView
-        expect(view.playPauseButtonImage, equals: .pause)
-    }
-
-    func testViewPlayPauseButtonImage2() {
-        state.isPlaying = false
-        let view = state.present() as! AudioBarView
-        expect(view.playPauseButtonImage, equals: .play)
-    }
-
-    func testViewPlaybackTime1() {
-        state.currentTime = nil
-        let view = state.present() as! AudioBarView
-        expect(view.playbackTime, equals: "")
-    }
-
-    func testViewPlaybackTime2() {
-        state.currentTime = 0
-        state.info = .init(duration: 1)
-        let view = state.present() as! AudioBarView
-        expect(view.playbackTime, equals: "-0:01")
-    }
-
-    func testViewPlaybackTime3() {
-        state.currentTime = 0
-        state.info = .init(duration: 61)
-        let view = state.present() as! AudioBarView
-        expect(view.playbackTime, equals: "-1:01")
-    }
-
-    func testViewPlaybackTime4() {
-        state.currentTime = 20
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.playbackTime, equals: "-0:40")
-    }
-
-    func testViewPlaybackDuration() {
-        state.info = .init(duration: .foo)
-        let view = state.present() as! AudioBarView
-        expect(view.playbackDuration, equals: .foo)
-    }
-
-    func testViewElapsedPlaybackTime1() {
-        state.currentTime = nil
-        let view = state.present() as! AudioBarView
-        expect(view.elapsedPlaybackTime, equals: 0)
-    }
-
-    func testViewElapsedPlaybackTime2() {
-        state.currentTime = .foo
-        let view = state.present() as! AudioBarView
-        expect(view.elapsedPlaybackTime, equals: .foo)
-    }
-
-    func testViewSeekIntervalWhenReadyToPlay() {
-        let view = state.present() as! AudioBarView
-        expect(view.seekInterval, equals: 15)
-    }
-
-    func testViewIsPlayCommandEnabled1() {
-        state.isPlaying = true
-        let view = state.present() as! AudioBarView
-        expect(view.isPlayCommandEnabled, equals: false)
-    }
-
-    func testViewIsPlayCommandEnabled2() {
-        state.isPlaying = false
-        state.currentTime = 0
-        state.info = .init(duration: 1)
-        let view = state.present() as! AudioBarView
-        expect(view.isPlayCommandEnabled, equals: true)
-    }
-
-    func testViewIsPlayCommandEnabled3() {
-        state.isPlaying = false
-        state.currentTime = 1
-        state.info = .init(duration: 1)
-        let view = state.present() as! AudioBarView
-        expect(view.isPlayCommandEnabled, equals: false)
-    }
-
-    func testViewIsPauseCommandEnabled1() {
-        state.isPlaying = false
-        let view = state.present() as! AudioBarView
-        expect(view.isPauseCommandEnabled, equals: false)
-    }
-
-    func testViewIsPauseCommandEnabled2() {
-        state.isPlaying = true
-        state.currentTime = 0
-        state.info = .init(duration: 1)
-        let view = state.present() as! AudioBarView
-        expect(view.isPauseCommandEnabled, equals: true)
-    }
-
-    func testViewIsPauseCommandEnabled3() {
-        state.isPlaying = true
-        state.currentTime = 1
-        state.info = .init(duration: 1)
-        let view = state.present() as! AudioBarView
-        expect(view.isPauseCommandEnabled, equals: false)
-    }
-
-    func testViewTrackName1() {
-        state.info = .init(title: nil)
-        let view = state.present() as! AudioBarView
-        expect(view.trackName, equals: nil)
-    }
-
-    func testViewTrackName2() {
-        state.info = .init(title: "A")
-        let view = state.present() as! AudioBarView
-        expect(view.trackName, equals: "A")
-    }
-
-    func testViewTrackName3() {
-        state.info = .init(title: "B")
-        let view = state.present() as! AudioBarView
-        expect(view.trackName, equals: "B")
-    }
-
-    func testViewArtistName1() {
-        state.info = .init(artist: nil)
-        let view = state.present() as! AudioBarView
-        expect(view.artistName, equals: nil)
-    }
-
-    func testViewArtistName2() {
-        state.info = .init(artist: "A")
-        let view = state.present() as! AudioBarView
-        expect(view.artistName, equals: "A")
-    }
-
-    func testViewArtistName3() {
-        state.info = .init(artist: "B")
-        let view = state.present() as! AudioBarView
-        expect(view.artistName, equals: "B")
-    }
-
-    func testViewAlbumName1() {
-        state.info = .init(album: nil)
-        let view = state.present() as! AudioBarView
-        expect(view.albumName, equals: nil)
-    }
-
-    func testViewAlbumName2() {
-        state.info = .init(album: "A")
-        let view = state.present() as! AudioBarView
-        expect(view.albumName, equals: "A")
-    }
-
-    func testViewAlbumName3() {
-        state.info = .init(album: "B")
-        let view = state.present() as! AudioBarView
-        expect(view.albumName, equals: "B")
-    }
-
-    func testViewArtworkData1() {
-        state.info = .init(artwork: nil)
-        let view = state.present() as! AudioBarView
-        expect(view.artworkData, equals: nil)
-    }
-
-    func testViewArtworkData2() {
-        state.info = .init(artwork: .init())
-        let view = state.present() as! AudioBarView
-        expect(view.artworkData, equals: .init())
-    }
-
-    func testViewIsLoadingIndicatorVisible1() {
-        state.isPlaying = true
-        state.currentTime = nil
-        let view = state.present() as! AudioBarView
-        expect(view.isLoadingIndicatorVisible, equals: true)
-    }
-
-    func testViewIsLoadingIndicatorVisible2() {
-        state.isPlaying = false
-        state.currentTime = nil
-        let view = state.present() as! AudioBarView
-        expect(view.isLoadingIndicatorVisible, equals: false)
-    }
-
-    func testViewIsLoadingIndicatorVisible3() {
-        state.isPlaying = true
-        state.currentTime = 0
-        let view = state.present() as! AudioBarView
-        expect(view.isLoadingIndicatorVisible, equals: false)
-    }
-
-    func testViewIsLoadingIndicatorVisible4() {
-        state.isPlaying = false
-        state.currentTime = 0
-        let view = state.present() as! AudioBarView
-        expect(view.isLoadingIndicatorVisible, equals: false)
-    }
-
-    func testViewIsSeekBackButtonEnabled1() {
-        state.currentTime = nil
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekBackButtonEnabled, equals: false)
-    }
-
-    func testViewIsSeekBackButtonEnabled2() {
-        state.currentTime = 0
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekBackButtonEnabled, equals: false)
-    }
-
-    func testViewIsSeekBackButtonEnabled3() {
-        state.currentTime = 1
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekBackButtonEnabled, equals: true)
-    }
-
-    func testViewIsSeekBackButtonEnabled4() {
-        state.currentTime = 2
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekBackButtonEnabled, equals: true)
-    }
-
-    func testViewIsSeekForwardButtonEnabled1() {
-        state.currentTime = nil
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekForwardButtonEnabled, equals: false)
-    }
-
-    func testViewIsSeekForwardButtonEnabled2() {
-        state.currentTime = 59
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekForwardButtonEnabled, equals: true)
-    }
-
-    func testViewIsSeekForwardButtonEnabled3() {
-        state.currentTime = 119
-        state.info = .init(duration: 120)
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekForwardButtonEnabled, equals: true)
-    }
-
-    func testViewIsSeekForwardButtonEnabled4() {
-        state.currentTime = 60
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.isSeekForwardButtonEnabled, equals: false)
-    }
-
-    func testViewIsPlayPauseButtonEnabled1() {
-        state.currentTime = nil
-        let view = state.present() as! AudioBarView
-        expect(view.isPlayPauseButtonEnabled, equals: true)
-    }
-
-    func testViewIsPlayPauseButtonEnabled2() {
-        state.currentTime = 60
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.isPlayPauseButtonEnabled, equals: false)
-    }
-
-    func testViewIsPlayPauseButtonEnabled3() {
-        state.currentTime = 59
-        state.info = .init(duration: 60)
-        let view = state.present() as! AudioBarView
-        expect(view.isPlayPauseButtonEnabled, equals: true)
-    }
-
-    func testViewIsPlayPauseButtonEnabled4() {
-        state.currentTime = 119
-        state.info = .init(duration: 120)
-        let view = state.present() as! AudioBarView
-        expect(view.isPlayPauseButtonEnabled, equals: true)
-    }
-
-}
